fix(reviews): reuse the real agent picker, unstretch the run button, stop repeating cancel
The reviewer picker was a plain Select listing bare harness ids, while project
settings already had a picker for the same choice with brand marks, proper
labels and install/auth state. That one is now shared: extracted as
ReviewerSelect and used in both places, so the Reviews tab shows "Claude Code"
with its logo and greys out anything not installed or authorised, instead of
"claude-code" as text. Its accessible name is a prop, since the same control now
names a project default in one place and a one-off override in the other.

The run action had flex-1, so dragging the inspector wider stretched the button
across the panel. It is a button, not a fill — it now keeps its size and the row
grows around it.

A cancelled run stores "cancelled by user" as its body, which the row rendered as
the review summary directly under a "Cancelled" badge, saying the same thing
twice. A terminated run's body is status, not findings, so it is no longer shown
as a summary.
